+++ plugin/src/net/groboclown/idea/p4ic/ui/config/P4SettingsPanel.java
@@ -17,7 +17,6 @@ package net.groboclown.idea.p4ic.ui.config;
 import com.intellij.openapi.Disposable;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.Disposer;
-import net.groboclown.idea.p4ic.config.ManualP4Config;
 import net.groboclown.idea.p4ic.config.P4ProjectConfigComponent;
 import net.groboclown.idea.p4ic.config.UserProjectPreferences;
 import net.groboclown.idea.p4ic.ui.config.props.ConfigStackPanel;